working on propagation hypotheses
